RAG/requirements.txt
streamlit
python-dotenv
langchain
langchain-openai
langchain-community
langchain-text-splitters
langchain-chroma
chromadb
pypdf
pillow
pytesseract
tiktoken
openai
groq
google-generativeai
pydantic
pytest
sentence-transformers
